requirements.txt
Flask
python-dotenv
markitdown
openai

websockets
torch
transformers
soundfile
numpy

